\name{plot.standardization}
\alias{plot.standardization}

\title{
Plot DDF
}
\description{
This function visualizes DDF summary indices and some other more things in the future.
}
\usage{
plot.standardization(x, type = "rmwsd", ...)
}

\arguments{
  \item{x}{
An object of class \code{standardization}.
}
  \item{type}{
Choose between different types of plots. Available: "rmwsd".
}
  \item{\dots}{
Further arguments for the \code{plot()} function.
}
}
\details{

There will be more plots ... soon.

}

\references{
Dorans, N. J. and Kullick, E. (1986). Demonstrating the utility of the standardization approach
to assessing unexpected differential item performance on the Scholastic Aptitude Test. Journal of 
Educational Measurement, 23, 355-368.

}
\author{
Manuel Reif
}



\seealso{
\code{\link{DDF}}
\code{\link{nrm}}
\code{\link{reshMG}}
}
\examples{

}

\keyword{DDF}

